requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
pandas>=2.2.0
requests>=2.31.0
seaborn>=0.13.0
tqdm>=4.66.0
